Добавил папку scripts и сложил туда скрипты. Добавил контейнер shopapi вместо ранее добавленного продюсера заглушки.
