+++ scripts/kafka-init.sh
