+++ shopapi/ca.crt
@@ -1,0 +1,25 @@
+-----BEGIN CERTIFICATE-----
+MIIEQTCCAymgAwIBAgIUKhirVBXqmMSEn0dAKNpjAHnvp3wwDQYJKoZIhvcNAQEL
+BQAwZTELMAkGA1UEBhMCUlUxDzANBgNVBAoMBllhbmRleDERMA8GA1UECwwIUHJh
+Y3RpY2UxDzANBgNVBAcMBk1vc2NvdzEhMB8GA1UEAwwYeWFuZGV4LXByYWN0aWNl
+LWthZmthLWNhMB4XDTI1MDMwMjA5NDYxN1oXDTI2MDMwMjA5NDYxN1owZTELMAkG
+A1UEBhMCUlUxDzANBgNVBAoMBllhbmRleDERMA8GA1UECwwIUHJhY3RpY2UxDzAN
+BgNVBAcMBk1vc2NvdzEhMB8GA1UEAwwYeWFuZGV4LXByYWN0aWNlLWthZmthLWNh
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA3DLyy1OMou0gyn+MYUyV
++123I3RU+FvQevR/qBV7A0x2FszObZPxkpxBLHz/SHpBvIY9v2iRGQYTnjoPInrM
+OSwT9WcwPpgUzDYouNed/TtAMryQ/986EzVOBxcgh+pj6gk8KxfP4KD536zUM3YC
+BSu88WMTMa2dMfzogn5P5RQnESYwUbdyvEl9gxQ0k5IcJFa9bGCUurXdaBMxXse/
+7/22TwK0T3kneXtpmkiwLq62Q32lR2R7bifv1AzOLyr146wRTjFTJE1hwtMTzI4T
+wdaTBPI2I27YpFYinKEkCCe/Sh0jdfCZrOUJGeoyQzOQ5IBk2RoD1cdt4owuvFG0
+QQIDAQABo4HoMIHlMB0GA1UdDgQWBBTn+Fb8wZmjCG/F1y6L+rFspno+pTAPBgNV
+HRMBAf8EBTADAQH/MIGiBgNVHSMEgZowgZeAFOf4VvzBmaMIb8XXLov6sWymej6l
+oWmkZzBlMQswCQYDVQQGEwJSVTEPMA0GA1UECgwGWWFuZGV4MREwDwYDVQQLDAhQ
+cmFjdGljZTEPMA0GA1UEBwwGTW9zY293MSEwHwYDVQQDDBh5YW5kZXgtcHJhY3Rp
+Y2Uta2Fma2EtY2GCFCoYq1QV6pjEhJ9HQCjaYwB576d8MA4GA1UdDwEB/wQEAwIB
+BjANBgkqhkiG9w0BAQsFAAOCAQEAMGk7/0mrRutOLmaSR6dw63OHnGmpDqfys3tf
+jG5Mp64Xc5Ilkf2T7/UZdUpFVqQn57nTLCdt6FCm4op+sjQm0X291P/aFp7UKx/9
+2valPlEMHFOlQvYveipdux/jloAhjmmbpAmLBNV2Nood9+lWpaYS7/M2UPvGtfYG
+CYIx14IY8Rwbdwbr2Mr0M9JjwXQbaNSzW/tiUj91VanSxJERQRyCe2XlD+TFIIAi
+T+8sTxRmiSN1PRSJ6dKwmzTVK2xTw9k8b4oNfdpf2f4ZZUAnWigfqM6ZkUZUIEZF
+GeojJa05aaEWTN5kC+VwSaFqozP7p7axnRADpfGPO8J/A073+Q==
+-----END CERTIFICATE-----
