+++ shopapi/kafka-1.crt
@@ -1,0 +1,25 @@
+-----BEGIN CERTIFICATE-----
+MIIENDCCAxygAwIBAgIUK4WAYwHuwcfezJeiwsOKVHFOFocwDQYJKoZIhvcNAQEL
+BQAwZTELMAkGA1UEBhMCUlUxDzANBgNVBAoMBllhbmRleDERMA8GA1UECwwIUHJh
+Y3RpY2UxDzANBgNVBAcMBk1vc2NvdzEhMB8GA1UEAwwYeWFuZGV4LXByYWN0aWNl
+LWthZmthLWNhMB4XDTI1MDMwMzIwMDkzOFoXDTM1MDMwMTIwMDkzOFowZDELMAkG
+A1UEBhMCUlUxDzANBgNVBAoMBllhbmRleDERMA8GA1UECwwIUHJhY3RpY2UxDzAN
+BgNVBAcMBk1vc2NvdzEgMB4GA1UEAwwXeWFuZGV4LXByYWN0aWNlLWthZmthLTEw
+ggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDGF9Y5I9Dqh0golvDUns2M
+lFnwohvNnmE0coSW7cVYHquKZUJj+6Iyhpc+aIsqmyE60wTYVGKrSvFS5vCtANle
+aacEHL9Li0MenMEv5+YS8tJnxGm0mEZzW0flrqAEnkdbIwJw79BG412S0ximVi0X
+Mw7Hc7OQefnXFw5x+y++e5uBXhaVQ1B7LjE8l5/lyjJq3f9OOnuHBtF05Gxzo6+w
+SuTxb8T/8pe6I2Dv0vO5wyDvlObAerNG6m8cMHUoSnKcaHbsigVYx7CpNI02pE4k
+Mf/ZkrEaD22wjf2j1cAOpLb18SzLF7+hw4qKv/tA+l5eAHlOFsff9xtYInMc42mP
+AgMBAAGjgdwwgdkwHQYDVR0OBBYEFEm5JYsrnRQo/uXrppA3GCkl6YeRMAkGA1Ud
+EwQCMAAwLAYJYIZIAYb4QgENBB8WHU9wZW5TU0wgR2VuZXJhdGVkIENlcnRpZmlj
+YXRlMA4GA1UdDwEB/wQEAwIFoDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUH
+AwIwLwYDVR0RBCgwJoIHa2Fma2EtMYIQa2Fma2EtMS1leHRlcm5hbIIJbG9jYWxo
+b3N0MB8GA1UdIwQYMBaAFOf4VvzBmaMIb8XXLov6sWymej6lMA0GCSqGSIb3DQEB
+CwUAA4IBAQAhRXn48ly6UP/dU3/3wmtbWDL5Iydu+NhLNVG/uJQ1hX5UliyMGxUy
+NT7dc4asAwqjdiKks9mlicRPqOISsWsWdBWeEASvzclm9T9D5RbnyEYeWX6TwtBi
+XvkGaxiA9Q4qv+xM05nJMRmt2gubJR1rWKJaoQMX011ixeO4aVUkihBI/vCxzlLd
+3yX4xqiKSIll85vxQEXrbr0czOhSFcMime1Aeb952phnYrxZtZDMyOvgS6wG/xgG
+7GoMEM750bLqXnchBemK+mJK2wZ5uq/vpOvUAfQPuD6ry1VVX7o5h9w9zcTDgOZZ
+o8HMPPWBXoxSmYWUi4YAjVr6MHHn9d+A
+-----END CERTIFICATE-----
