+++ shopapi/kafka-1.key
@@ -1,0 +1,28 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDGF9Y5I9Dqh0go
+lvDUns2MlFnwohvNnmE0coSW7cVYHquKZUJj+6Iyhpc+aIsqmyE60wTYVGKrSvFS
+5vCtANleaacEHL9Li0MenMEv5+YS8tJnxGm0mEZzW0flrqAEnkdbIwJw79BG412S
+0ximVi0XMw7Hc7OQefnXFw5x+y++e5uBXhaVQ1B7LjE8l5/lyjJq3f9OOnuHBtF0
+5Gxzo6+wSuTxb8T/8pe6I2Dv0vO5wyDvlObAerNG6m8cMHUoSnKcaHbsigVYx7Cp
+NI02pE4kMf/ZkrEaD22wjf2j1cAOpLb18SzLF7+hw4qKv/tA+l5eAHlOFsff9xtY
+InMc42mPAgMBAAECggEABBXGwavm6Yc7qjaepcvWklEW37sAFsfu2j2IVXJ20N/l
+CrgW2+f7Cs6yvMM8J3dGgqeKYImw9TZY7qw7h13Onb9RBekv2rd2/WEfLMU7/eUL
+CL/IrEQ9o7BctbRZsW6wKwV4TersISIcMmRvDDMXv/Q6GEvfBpo1WEWt8xTKgcOS
+gVNHPSyZsCNAbED+dmO7ID34c8BDNWDpEMeNcuEA8LgYRt+tXhjU+gMcLWEQnCLv
+I8a/hgD8WsqOVGcQVwaYd3qK446w+iRZrhmn4/Z2snlodwLUNlvMM2LBcX1+Tasg
+ZO2ANqU3IYMreo05rdKiB4s4rnEwxbogECKfOjidyQKBgQDnQ0GyzBb/F7kHZ3hp
+LpuSNbzwbatZYRJTE61BI0JvllxX0BRu4i6cuqAGvCPaBZr4bnb3EFlV+mTxxMYB
+rc3GVQ1t5rQqhHdHspC8MAO85RED4mFyAsoWh7pGwrV1Je0JII513c1C3Ne/2p0I
+/B4zRbqkodQE1EbyoX9qWNQsdwKBgQDbSEks3P9zoNdFZFasLXARM4A0VxhH0OIh
+bwVPCFCdodVb+2MOjwAbaxY2Bl3vtZG1A3VkIhBwMDhfhAtRFPh1tkI/FUIBsR1l
+LzZtv27oxR96/Ij/+pJsPraTgZN/obLWQ6W+AKANcTxL7LZFUfcwRTZZtXjM6Tu0
+vVOBE6UpqQKBgE20zq2KYCvEx9pSDSaij4mYXu0WnJ3CvUqG8Q9TDq66AjtLYqJD
+g9AL8Vtx+kX23jSdwYGQz8PRB/wo1zOXN+5tg76q51zl98JyoKlrJMZi3c3FUmLh
+xw+8+X1vTVOYs8X+V4AJUEAoPqJrjZqIRImL44QQEFKzFUfPtiWAxYJPAoGBAJcY
+egOFzCd6fpr8BpThQkU3/XETSlHB5u8utrO8gP/9ES0uTRUtB9/crBNR08xmih32
+Yh87MaR8NXvrWnjmQEWr7n3VQgQ0bJLaM2Dv9ZZCrexqsmPf3qDjnZ6KPH5HhxkI
+x93iRGa+zdNrKjUdw3vpVQf/eVuIonFpu25HgRjxAoGAPIPom0AywsUsXwQTb1aP
+9W+09MYhD64vwWsJdc8EdOXfk7bXq+vYA8nS1GboRXjSqp5X0yPdz/MEK6xOfCiq
+qrjjkTz0edcT63Sq53K0nDdoNpMQv/rY/VglcG8OEeAYIjHk9rzxZ5dl5CHGKzw4
+HjPk8eIQhSjH/8acyRfj6Zc=
+-----END PRIVATE KEY-----
